requirements.txt
pandas
numpy
gensim
spacy
pyLDAvis
matplotlib
ipython
httpx
mysql-connector-python==8.0.33
python-dotenv==1.0.0
tqdm

